[fix] CreateDietResponseDTO DietFoodForCreateDietDTO.intake 타입 Int → Double
API 서버가 소수 섭취량을 반환할 경우 디코딩 실패하는 버그 수정

+++ BalanceEat/Data/DTOs/CreateDietResponseDTO.swift
@@ -28,7 +28,7 @@ struct DietFoodForCreateDietDTO: Codable {
     let id: Int
     let foodId: Int
     let foodName: String
-    let intake: Int
+    let intake: Double
     let nutrition: FoodNutritionForCreateDietDTO
 }
